Splitted Flatten Syntax in two.
Conflicts:
	src/javasources/KTool/src/org/kframework/compile/transformers/FlattenSyntax.java
